[NuGet] Fix package compatibility check when multiple solutions open.
Previously if multiple solutions were open in the workspace then
changing the project's target framework would not work in a solution
after one of the solutions are closed. All the event handlers were
removed for all projects. Also when the target framework was changed
in the project file outside of the IDE the package compatibility check
would not be run. This was because event handlers for the solution's
SolutionItemAdded event were removed.